pyperf3/requirements.txt
python-decouple
httpx
